chore: Update keyword search functionality in Home page
This commit updates the Home page to include a keyword search functionality. The user can enter a keyword in the input field and click the "Submit" button to search for posts containing that keyword. The search results are displayed in a list format below the input field. This enhancement improves the user experience by allowing them to easily find relevant posts based on their search criteria.