chore: Update Admin page to fetch and display users
This commit updates the Admin page to fetch and display users. It adds the necessary code to fetch the list of users from the server and store them in the `users` state. The fetched users are then displayed in a tab panel using the `JSON.stringify` function. This enhancement improves the functionality of the Admin page by allowing the user to view and manage the list of users.